Initialize Rollbar
This will initialize Rollbar. I'm *not* a Laravel user, so this is my best guess at how it would work.

The sooner you can initialize Rollbar the sooner you can catch errors. For instance, this will not catch any errors that happen *before* init is called.

I also don't know much PHP. Here are the assumptions I built this on:

 1. It assumes that `Config::get` can request the whole config file, and that it will return the associative array if you do.
 2. It assumes that `NULL` is not truthy.